move stories from old-docs to docs
